requirements.txt
Django>=4.2,<5.0
django-crispy-forms>=2.0
crispy-bootstrap5>=0.7
reportlab>=4.0
Pillow>=10.0
python-dateutil>=2.8
